feat: add Vercel KV

+++ actions/deleteTodoCollection.ts
@@ -2,14 +2,24 @@
 
 import { connectMongoDB } from '@/libs/mongodb';
 import TodoCollection from '@/models/todoCollection';
+import { kv } from '@vercel/kv';
 
 export default async function deleteTodoCollection(todoCollectionId: string) {
 	try {
-		await connectMongoDB();
+		if (process.env.USE_MONGODB === 'true') {
+			await connectMongoDB();
+			await TodoCollection.findOneAndDelete({ id: todoCollectionId });
+		}
 
-		await TodoCollection.findOneAndDelete({ id: todoCollectionId });
+		if (process.env.USE_VERCEL_KV === 'true') {
+			await kv.del(todoCollectionId);
+		}
 
-		return { message: 'TODO collection is deleted.', error: null };
+		if (!process.env.USE_MONGODB && !process.env.USE_VERCEL_KV) {
+			throw new Error('No Storage Provider Specified.');
+		}
+
+		return { message: 'TODO Collection Deleted', error: null };
 	} catch (e) {
 		console.error(e);
